pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "A beautiful CLI for getting weather information using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "beautifulsoup4",
    "selenium",
    "lxml",
    "httpx",
    "python-dotenv",
    "click>=8.0.0",
    "langchain-openai",
    "langgraph",
    "markdownify",
]

[project.scripts]
weather = "src.main:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "responses",
    "black",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
